requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.12.0
    # via
    #   google-genai
    #   httpx
    #   openai
    #   watchfiles
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
backoff==2.2.1
    # via posthog
bcrypt==5.0.0
    # via chromadb
build==1.3.0
    # via chromadb
cachetools==6.2.2
    # via google-auth
certifi==2025.11.12
    # via
    #   httpcore
    #   httpx
    #   kubernetes
    #   requests
charset-normalizer==3.4.4
    # via requests
chromadb==1.3.6
    # via langchain-chroma
click==8.3.1
    # via
    #   typer
    #   typer-slim
    #   uvicorn
colorama==0.4.6
    # via
    #   build
    #   click
    #   tqdm
    #   uvicorn
coloredlogs==15.0.1
    # via onnxruntime
distro==1.9.0
    # via
    #   google-genai
    #   openai
    #   posthog
dotenv==0.9.9
    # via simple-rag-agents (pyproject.toml)
durationpy==0.10
    # via kubernetes
filelock==3.20.0
    # via huggingface-hub
flatbuffers==25.9.23
    # via onnxruntime
fsspec==2025.12.0
    # via huggingface-hub
google-auth==2.43.0
    # via
    #   google-genai
    #   kubernetes
google-genai==1.55.0
    # via simple-rag-agents (pyproject.toml)
googleapis-common-protos==1.72.0
    # via opentelemetry-exporter-otlp-proto-grpc
grpcio==1.76.0
    # via
    #   chromadb
    #   opentelemetry-exporter-otlp-proto-grpc
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
hf-xet==1.2.0
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via
    #   chromadb
    #   google-genai
    #   huggingface-hub
    #   langsmith
    #   openai
huggingface-hub==1.2.3
    # via tokenizers
humanfriendly==10.0
    # via coloredlogs
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
importlib-metadata==8.7.0
    # via opentelemetry-api
importlib-resources==6.5.2
    # via chromadb
jiter==0.12.0
    # via openai
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
jsonschema==4.25.1
    # via chromadb
jsonschema-specifications==2025.9.1
    # via jsonschema
kubernetes==33.1.0
    # via chromadb
langchain-chroma==1.1.0
    # via simple-rag-agents (pyproject.toml)
langchain-core==1.2.0
    # via
    #   langchain-chroma
    #   langchain-openai
langchain-openai==1.1.2
    # via simple-rag-agents (pyproject.toml)
langsmith==0.4.59
    # via langchain-core
markdown-it-py==4.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
mmh3==5.2.0
    # via chromadb
mpmath==1.3.0
    # via sympy
numpy==2.3.5
    # via
    #   chromadb
    #   langchain-chroma
    #   onnxruntime
    #   pandas
oauthlib==3.3.1
    # via
    #   kubernetes
    #   requests-oauthlib
onnxruntime==1.23.2
    # via chromadb
openai==2.11.0
    # via
    #   simple-rag-agents (pyproject.toml)
    #   langchain-openai
opentelemetry-api==1.39.1
    # via
    #   chromadb
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-exporter-otlp-proto-common==1.39.1
    # via opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-grpc==1.39.1
    # via chromadb
opentelemetry-proto==1.39.1
    # via
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-grpc
opentelemetry-sdk==1.39.1
    # via
    #   chromadb
    #   opentelemetry-exporter-otlp-proto-grpc
opentelemetry-semantic-conventions==0.60b1
    # via opentelemetry-sdk
orjson==3.11.5
    # via
    #   chromadb
    #   langsmith
overrides==7.7.0
    # via chromadb
packaging==25.0
    # via
    #   build
    #   huggingface-hub
    #   langchain-core
    #   langsmith
    #   onnxruntime
pandas==2.3.3
    # via simple-rag-agents (pyproject.toml)
posthog==5.4.0
    # via chromadb
protobuf==6.33.2
    # via
    #   googleapis-common-protos
    #   onnxruntime
    #   opentelemetry-proto
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pybase64==1.4.3
    # via chromadb
pydantic==2.12.5
    # via
    #   chromadb
    #   google-genai
    #   langchain-core
    #   langsmith
    #   openai
pydantic-core==2.41.5
    # via pydantic
pygments==2.19.2
    # via rich
pypika==0.48.9
    # via chromadb
pyproject-hooks==1.2.0
    # via build
pyreadline3==3.5.4
    # via humanfriendly
python-dateutil==2.9.0.post0
    # via
    #   kubernetes
    #   pandas
    #   posthog
python-dotenv==1.2.1
    # via
    #   dotenv
    #   uvicorn
pytz==2025.2
    # via pandas
pyyaml==6.0.3
    # via
    #   chromadb
    #   huggingface-hub
    #   kubernetes
    #   langchain-core
    #   uvicorn
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2025.11.3
    # via tiktoken
requests==2.32.5
    # via
    #   google-auth
    #   google-genai
    #   kubernetes
    #   langsmith
    #   posthog
    #   requests-oauthlib
    #   requests-toolbelt
    #   tiktoken
requests-oauthlib==2.0.0
    # via kubernetes
requests-toolbelt==1.0.0
    # via langsmith
rich==14.2.0
    # via
    #   chromadb
    #   typer
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
rsa==4.9.1
    # via google-auth
shellingham==1.5.4
    # via
    #   huggingface-hub
    #   typer
six==1.17.0
    # via
    #   kubernetes
    #   posthog
    #   python-dateutil
sniffio==1.3.1
    # via
    #   google-genai
    #   openai
sympy==1.14.0
    # via onnxruntime
tenacity==9.1.2
    # via
    #   chromadb
    #   google-genai
    #   langchain-core
tiktoken==0.12.0
    # via langchain-openai
tokenizers==0.22.1
    # via chromadb
tqdm==4.67.1
    # via
    #   simple-rag-agents (pyproject.toml)
    #   chromadb
    #   huggingface-hub
    #   openai
typer==0.20.0
    # via chromadb
typer-slim==0.20.0
    # via huggingface-hub
typing-extensions==4.15.0
    # via
    #   anyio
    #   chromadb
    #   google-genai
    #   grpcio
    #   huggingface-hub
    #   langchain-core
    #   openai
    #   opentelemetry-api
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   pydantic
    #   pydantic-core
    #   referencing
    #   typer
    #   typer-slim
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==2.6.2
    # via
    #   kubernetes
    #   requests
uuid-utils==0.12.0
    # via
    #   langchain-core
    #   langsmith
uvicorn==0.38.0
    # via chromadb
watchfiles==1.1.1
    # via uvicorn
websocket-client==1.9.0
    # via kubernetes
websockets==15.0.1
    # via
    #   google-genai
    #   uvicorn
zipp==3.23.0
    # via importlib-metadata
zstandard==0.25.0
    # via langsmith
